+++ ast/utils.go
@@ -15,3 +15,19 @@ func findSlotStmtIndex(stmts []Statement, slotName string) int {
 
 	return -1
 }
+
+func findDuplicateSlot(slots []*SlotStmt) (string, int) {
+	counts := make(map[string]int)
+
+	for _, slot := range slots {
+		counts[slot.Name.Value]++
+	}
+
+	for name, count := range counts {
+		if count > 1 {
+			return name, count
+		}
+	}
+
+	return "", 0
+}